Fix build: LineItem.unitQty expects Int, not Long (#139)
compileDebugKotlin failed on a real build: LineItem.unitQty's setter takes
Int (confirmed by the compiler error), not Long. Drop the incorrect
.toLong() conversion in both places the Quantity widget writes it back
(OverlayActivity's register overlay, OrderDetailFragment's order-detail
screen) - quantity is already an Int coming from
ItemNoteListener.onNoteSaved.

+++ app/src/main/java/com/orderMate/fragment/orderDetail/OrderDetailFragment.kt
@@ -1177,7 +1177,7 @@ class OrderDetailFragment : Fragment(), IOrderItemClickListener, ILineItemUpdate
                                     if (lineItem?.item?.id == itemId) {
                                         android.util.Log.d("ItemNoteReceivedDebug", "Setting note on lineItem: ${lineItem.id}")
                                         lineItem.note = note
-                                        lineItem.unitQty = quantity.toLong()
+                                        lineItem.unitQty = quantity
                                     }
                                 }
